初始化

+++ src-tauri/src/infra/time.rs
@@ -48,7 +48,7 @@ fn civil_from_days(days: i64) -> CivilDate {
     // 把纪元挪到 0000-03-01，让闰年落在周期末尾
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
